+++ CHANGES.rst
@@ -1,8 +1,14 @@
 Changelog for Rapid Photo Downloader
 ====================================
 
-0.9.15 (2019-06-23)
--------------------
+0.9.15b1 (2019-06-24)
+---------------------
+
+ - Fixed bug #1829145 where Rapid Photo Downloader could no longer access
+   cameras, smartphones and tablets because other applications using GIO / GVFS
+   (such as a file manager like Gnome Files) had gained exlusive access over
+   them. GIO / Glib changed the way paths were generated for cameras and phones
+   in a way that was incompatible with libgphoto2's port nomenclature.
 
  - Fix bug #1818280: sqlite3 database is locked while adding thumbnails.